Move missing blobs check into remote_node.
